requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gunicorn==21.2.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
python-dotenv==1.0.1
python-multipart==0.0.6
python-docx==1.1.2
lxml==4.9.3
PyJWT==2.8.0